refact: move scattered sub-class of Feature to base/feature
